+++ src/model/food_bag.rs
@@ -2,12 +2,12 @@ use crate::model::food::Food;
 
 pub struct FoodBag {
     items: Vec<Food>,
-    total_calories: u16,
+    total_calories: u32,
 }
 
 impl FoodBag {
     pub fn new(items: Vec<Food>) -> Self {
-        let mut total_calories: u16 = 0;
+        let mut total_calories: u32 = 0;
         for food in items.iter() {
             total_calories = total_calories + food.calories;
         }
@@ -18,8 +18,8 @@ impl FoodBag {
         }
     }
 
-    pub fn total_calories(&self) -> u16 {
-        let mut total: u16 = 0;
+    pub fn total_calories(&self) -> u32 {
+        let mut total: u32 = 0;
         for food in self.items.iter() {
             total = total + food.calories;
         }
